% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A Welcome function}
\usage{
hello()
}
\value{
prints a very welcome
}
\description{
Just be friendly!
}
\note{
This function is a bit boring but that is ok.
}
\examples{
hello()
}
\seealso{
[Arithmetic] for an easier way.
}
