% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{Some dummy data}
\format{
## `my_data`
A data frame containing 6 rows and 2 columns:
\describe{
  \item{X}{Covariate}
  \item{Y}{Endpoint}
  ...
}
}
\usage{
my_data
}
\description{
Some example data that could help
}
\keyword{datasets}
