% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sum_factorial.R
\name{my_sum_factorial}
\alias{my_sum_factorial}
\title{My Summation Function of 2 factorials}
\usage{
my_sum_factorial(x, y)
}
\arguments{
\item{x}{first summand.}

\item{y}{second summand.}
}
\value{
The sum of `x!` and `y!`.
}
\description{
This is my first function and it first takes the factorial of two numbers and then sums two factorials.
}
\note{
This function is a bit boring but that is ok.
}
\examples{
my_sum_factorial(1, 2)
}
\seealso{
[Arithmetic] for an easier way.
}
